+++ src/Logary/Constants.fs
@@ -37,28 +37,41 @@ module KnownLiterals =
 
 /// Time calculation constants
 module Constants =
+  /// BCL ticks. Not the same as Stopwatch ticks.
   [<Literal>]
   let SecondsPerTick = 0.0000001
+
   [<Literal>]
   let MillisPerTick = 0.0001
   [<Literal>]
   let MicrosPerTick = 0.1
+
   [<Literal>]
   let NanosPerTick = 100L
   [<Literal>]
-  let NanosPerMicro = 1000L
+  let NanosPerMicro = 1_000L
   [<Literal>]
-  let NanosPerMilli = 1000000L
+  let NanosPerMilli = 1_000_000L
   [<Literal>]
-  let NanosPerSecond = 1000000000L
+  let NanosPerSecond = 1_000_000_000L
   [<Literal>]
-  let NanosPerMinute = 60000000000L
+  let NanosPerMinute = 60_000_000_000L
 
   [<Literal>]
-  let TicksPerMinute = 600000000L
+  let MicrosPerSecond = 1_000_000L
+
   [<Literal>]
-  let TicksPerSecond = 10000000L
+  let MillisPerSecond = 1_000L
+
+  /// BCL ticks. Not the same as Stopwatch ticks.
+  [<Literal>]
+  let TicksPerMinute = 600_000_000L
+  /// BCL ticks. Not the same as Stopwatch ticks.
+  [<Literal>]
+  let TicksPerSecond = 10_000_000L
+  /// BCL ticks. Not the same as Stopwatch ticks.
   [<Literal>]
   let TicksPerMilli = 10000L
+  /// BCL ticks. Not the same as Stopwatch ticks.
   [<Literal>]
   let TicksPerMicro = 10L