+++ src/Logary/Extensions/NodaTime.Instant.fs
@@ -3,5 +3,6 @@ namespace Logary
 open NodaTime
 
 module Instant =
+
   let ofEpoch (e: EpochNanoSeconds) =
     NodaTime.Instant.FromUnixTimeTicks(e / Constants.NanosPerTick)